+++ tasks/seq/oturin_a_max_values_by_rows_matrix_seq/func_tests/main.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 
-#include <vector>
 #include <numeric>
+#include <vector>
 
 #include "seq/oturin_a_max_values_by_rows_matrix_seq/include/ops_seq.hpp"
 
@@ -10,13 +10,12 @@ TEST(Sequential, Test_Max_5_5) {
   size_t m = 5;
 
   // Create data
-  std::vector<int> in(n*m);
+  std::vector<int> in(n * m);
   std::vector<int> out(m, 0);
   std::vector<int> maxes(m);
 
   std::iota(std::begin(in), std::end(in), 1);
-  for (int i = 0; i < m; i++) 
-    maxes[i] = (i+1)*n;
+  for (size_t i = 0; i < m; i++) maxes[i] = (i + 1) * n;
 
   // Create TaskData
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
@@ -39,13 +38,12 @@ TEST(Sequential, Test_Max_10_5) {
   size_t m = 5;
 
   // Create data
-  std::vector<int> in(n*m);
+  std::vector<int> in(n * m);
   std::vector<int> out(m, 0);
   std::vector<int> maxes(m);
 
   std::iota(std::begin(in), std::end(in), 1);
-  for (int i = 0; i < m; i++) 
-    maxes[i] = (i+1)*n;
+  for (size_t i = 0; i < m; i++) maxes[i] = (i + 1) * n;
 
   // Create TaskData
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
@@ -67,13 +65,12 @@ TEST(Sequential, Test_Max_5_10) {
   size_t n = 5, m = 10;
 
   // Create data
-  std::vector<int> in(n*m);
+  std::vector<int> in(n * m);
   std::vector<int> out(m, 0);
   std::vector<int> maxes(m);
 
   std::iota(std::begin(in), std::end(in), 1);
-  for (int i = 0; i < m; i++) 
-    maxes[i] = (i+1)*n;
+  for (size_t i = 0; i < m; i++) maxes[i] = (i + 1) * n;
 
   // Create TaskData
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
